[FIX] 슬랙 모듈 분리
